requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
scrapy>=2.11.0

# Calendar & Datetime
icalendar>=5.0.11
caldav>=1.3.9
python-dateutil>=2.8.2
pytz>=2023.3

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
schedule>=1.2.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
pylint>=3.0.0
